Addresses webauthn#1643
This is a proposal for the I18N WG to consider in our teleconference of
2021-09-09 and is not yet ready for webauthn.

Basically this changes the sequence to use ^^^ as an introducer,
followed optionally by a lowercase language tag and/or a direction
sequence. The direction sequence is introduced with an underscore (\_)
and consists of either `ltr` or `rtl`. Other text was adjusted as
appropriate to these changes. The resulting sequence is consistent with
JSON-LD's representation, only it leaves off the `i18n:` namespace
identifier.

Authoring guidance is probably warranted, particularly about using short
language tags. I did not include a closing marker to signal truncation
despite webauthn's request that we do so.